Move CodeEditor into lib
